+++ docs/release-notes.md
@@ -27,6 +27,12 @@ of `zizmor`.
 
 ### Enhancements
 
+* Recommend `gh issue edit --add-label` / `gh pr edit --add-label` as a replacement for
+  @actions-ecosystem/action-add-labels in [superfluous-actions]
+
+* Recommend `gh issue edit --remove-label` / `gh pr edit --remove-label` as a replacement for
+  @actions-ecosystem/action-remove-labels in [superfluous-actions]
+
 * @tibdex/github-app-token is now recognized as an archived action by
   [archived-uses] (#1910)